Create gitignore from CMS example
This is the gitignore used by the CMS project. It is entirely possible that it is missing some things or has some extra stuff we don't need. Please review it and decide if there is anything that needs to be changed before approving the merge. If you've approved it and I haven't merged it yet, just send me an @ on discord. Thanks